requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
markdown>=3.5.0
